Removed year from isTollFreeDate()
Reason : This commit was made to remove the inflexibility from the method and make the system compatible to all years
Fault found : The system only checks if it's a holiday for year 2013
Suggested fix : Since we take the current year there is no point to validate the year. We can remove the year and compare the date with the pre-defined holidays to check if it's a holiday instead.
The inevitable : Since holidays are different from a country to another country, we will have to modify the pre-defined holidays according to the local holidays.